+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Changelog
 Unreleased
 ----------
 - Add class TechnicalAudioMetadata for creating technical audio metadata in AudioMD standard
+- Add class TechnicalVideoMetadata for creating technical video metadata in VideoMD standard
 
 All notable changes to this project will be documented in this file.